+++ eventmesh-storage/eventmesh-storage-kafka/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.connector.StorageResourceService
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-rabbitmq=org.apache.eventmesh.storage.rabbitmq.connector.RabbitmqStorageResourceService
+kafka=org.apache.eventmesh.storage.kafka.storage.StorageResourceServiceKafkaImpl
