+++ eventmesh-storage/eventmesh-storage-rabbitmq/src/main/java/org/apache/eventmesh/storage/rabbitmq/storage/RabbitmqStorageResourceService.java
@@ -8,18 +8,18 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
  */
 
-package org.apache.eventmesh.storage.knative.connector;
+package org.apache.eventmesh.storage.rabbitmq.storage;
 
 import org.apache.eventmesh.api.storage.StorageResourceService;
 
-public class StorageResourceServiceKnativeImpl implements StorageResourceService {
+public class RabbitmqStorageResourceService implements StorageResourceService {
 
     @Override
     public void init() throws Exception {
