+++ eventmesh-storage/eventmesh-storage-rabbitmq/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.connector.StorageResourceService
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-kafka=org.apache.eventmesh.storage.kafka.connector.StorageResourceServiceKafkaImpl
+rabbitmq=org.apache.eventmesh.storage.rabbitmq.storage.RabbitmqStorageResourceService
