+++ eventmesh-storage/eventmesh-storage-redis/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.connector.StorageResourceService
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-redis=org.apache.eventmesh.storage.redis.connector.RedisStorageResourceService
+redis=org.apache.eventmesh.storage.redis.storage.RedisStorageResourceService
