+++ eventmesh-storage/eventmesh-storage-redis/src/test/java/org/apache/eventmesh/storage/redis/storage/UnitTest.java
@@ -15,7 +15,7 @@
  *  limitations under the License.
  */
 
-package org.apache.eventmesh.storage.redis.connector;
+package org.apache.eventmesh.storage.redis.storage;
 
 import org.apache.eventmesh.api.EventMeshAction;
 import org.apache.eventmesh.api.SendCallback;
